Updates signup middleware tests
